Add isValid bit to Touchpoint class
Summary:
Every time we run an attribution rule, we have to calculate `isNewTouchpointValid`, which is a greater-than-or-equal comparison between `emp::Integers`. My hunch was that it'd be faster to pre-compute this and replace it with just an `emp::Bit` check.

Indeed it looks like this does improve performance slightly, with a 6% reduction in runtime, a 0.8% reduction in network traffic, and a 1.2% reduction in cost. (Though I'd take these numbers with a grain of salt since there is variability every time we run the perf script.)

Reviewed By: gorel

Differential Revision: D30123838

fbshipit-source-id: c7b36a8eefbf729233218459d718de1a1d03d3e5