Support lists in gen_config --from argument
Summary:
## What
Support auto-fill for list-based arguments in a more elegant way.

## Why
Support auto-fill for list-based arguments in a more elegant way.t the user writes something like `subnet-123, subnet-456` inside of the TODO (but the brackets make it clear that the argument is a list).

When trying to autofill, by default we get something dumb like `[[subnet-123, subnet-456]]` because it doesn't realize we already "helpfully" placed the brackets there.

There's still some weirdness if someone does something like providing a prompt that looks like `[abc, TODO, ghi]` but I really can't think of _why_ we would do that in the first place. It seems like awful UX to have a positional-based argument that *must* be taken from the user.

In any case, this update script has a narrow use-case of being done by manually onboarded clients or in our own Cloudbridge deployment, so we should be able to handle any esoteric cases like that ourselves.

Reviewed By: peking2

Differential Revision: D30297208

fbshipit-source-id: 4cab737b74d4dda66aff045adb111c8c804740ed