Decouple instance_base from fbpcp
Summary:
According to MVCS design pattern, the entity should be a plain object which represents data only. For persistence, it's the responsibility of a repository. A mapper is responsible for converting data from one format to the other.

It seemed instance_base was introduced for the convenience of implementing a repository (D23430773). However it also introduced tight coupling.

It's ok for solutions to assume this coupling, but may not be good for assuming it in FBPCP. For example, it's tightly coupled with JSON now, but FBPCP shouldn't enforce the user to stick with JSON. Hence, we want to move it from FBPCP to FBPCS.

In FBPCP, the only class which depends on it is mpc_instance.py. I will take a look how to decouple it in the next diff.

Reviewed By: jrodal98

Differential Revision: D30354781

fbshipit-source-id: e20e0cf9fbaa07e8f5df2df0979bb915fe02cf8f