Make num_files_per_mpc_container required (#25)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/25

`num_files_per_mpc_container` is a piece of information that's required to complete a Private Computation. When creating a PA instance, it's already required ([code pointer](https://fburl.com/diffusion/r45hnlpr)) for the user to provide it. For PL, it's not required to be passed in from the user because it always uses a hard-coded value = 40. Either case, it makes sense to save this information in PrivateComputationInstance as a required attribute.

Reviewed By: gorel

Differential Revision: D30500380

fbshipit-source-id: bed7789390174cb3afbc4c070f07f5a371edd76f