Add docker-publish.yml file to publish onedocker image (#47)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/47

Process to publish the image:

1.Cron job is scheduled at Mon-Fri 17:00 UTC( 10:00AM PT)
2.Cron job will check if there are new commits in the past 24 hours. If no new commits, it wont' proceed.
3.If there are new commits(or manually trigger the workflow), it will skip new commits check and move onto next step to build a onedocker image, and tag the image with most recent commit_id(github.sha) and push the image to ghcr.io/facebookresearch/fbpcs/rc/onedocker

Why using a /rc prefix?
rc/ondocker is for all the images built everyday; only after e2e tests passed, the image will be pushed to /onedocker; When setting up in this way, we can track the image availability in the long term(like XX days in a month, image is automatically built and passed E2E tests without human intervention)

This is tested on my own repo:
https://pxl.cl/1P4db

4.Before E2E tests in place, it requires on-call manually intervention to publish the image.

How to do it:  Once oncall finishes the weekly release, they can find the image which covers all changes passed the test, and promote that image as :latest. For example, if oncall is releasing bundles including the changes after daily image built, they will safely promote the image running on the same day; otherwise, they will choose an image from previous day.

How to manually approve: Find the workflows running for that day, there is a button called "push", once click on it, we can manually promote that images as :latest  and :main and move it to  ghcr.io/facebookresearch/fbpcs/onedocker

 Looks like
{F661282548}

I will use T100098035 to add other  on-call individuals to reviewers list

Reviewed By: corbantek, peking2, ajaybhargavb

Differential Revision: D30789072

fbshipit-source-id: b2bcc12f6c9c82af78958f519291178efe6aebe4