get rid of docker/terraform interpolation-expression only warning (#42)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/42

Previously we see the following message:
```Warning: Interpolation-only expressions are deprecated

  on glue.tf line 25, in resource "aws_iam_role_policy_attachment" "glue_service":
  25:   role       = "${aws_iam_role.glue_service_role.id}"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Template interpolation syntax is still used to construct strings from
expressions when the template includes multiple interpolation sequences or a
mixture of literal strings and interpolations. This deprecation applies only
to templates that consist entirely of a single interpolation sequence.
```
In this diff, we clean it up by changing
```
"${aws_iam_role.glue_service_role.id}"
```
to
```
aws_iam_role.glue_service_role.id
```

Reviewed By: peking2

Differential Revision: D30737116

fbshipit-source-id: c4db69a07de273cb4c620038adbccd05412c199d