Give user the option to provide game_type (#71)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/71

Pass a `game_type` to PL service for it to save in instance at instance creation time and to consume it during computation in D30978299. It's required in the PrivateLiftService interface. I'd love to make it required at the pl-coordinator and thrift service level as well, but I cannot do it without breaking one-command and pl graph api, so that's not addressed in this diff.

Intentionally not applying the same changes to PA because later in this stack I will rename PrivateLiftService to PrivateComputationService and make pa-coordinator use PrivateComputationService. Once that happens, I will add the game_type argument to pa-coordinator. PrivateAttributionService will be deleted.

Reviewed By: peking2, leegross

Differential Revision: D30856495

fbshipit-source-id: aad6becd56c63c504c085035cfc0db481d0ba006