Apply clang-format to OutputMetrics.hpp (#80)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/80

# What
* Simply applying `clang-format -i` to the OutputMetrics.hpp file to clean it up
* Otherwise this diff is a no-op

# Why
I started making changes (now moved to the next diff) and found a bunch of formatting in this file was off. I made this intermediate diff to purely apply formatting changes so it can be easy to review and separate the actual code changes being made in the *next* diff in the stack.

Reviewed By: corbantek, elliottlawrence

Differential Revision: D30413247

fbshipit-source-id: b9b921d4fdf99c7ab386727d5c5a42fadb982457