Remove --game=<game_name> arguments from PA-Coordinator (#91)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/91

This is a BE follow up from [4/n] D31106166 and [5/n] D31107234. For historical reasons, users are allowed to pass game_name arguments to pa-coordinator. In reality, because there's no more than one game for attribution, and no more than one game for shard aggregator, so in this diff we remove the game_name arguments.

Reviewed By: leegross

Differential Revision: D31116432

fbshipit-source-id: 16f4af347e08b78981ac290b2e4b3796e1846b7e