Add script for deploying shared PCE resources
Summary:
PCE shared resources are shared across a single region. So we only need to manually run this script once per region and the resources will be shared across all advertisers in that region.
Note: PCE service library will have the logic to understand shared PCE resources and it will automatically use it on the publisher side

Next: Add a wiki page with instructions on how to deploy shared resources

Differential Revision: D31322556

fbshipit-source-id: a1b94273e9417de704f58f13936240ad4824f087