Pass in hmac_key at PL instance creation time (#100)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/100

Remove the option to pass in hmac_key at the the id_match stage for PL. Instead, take it at create_instance and save it to instance. PA is already only taking it at create_instance and not at id_match. No Thrift layer change because the publisher side do not need the hmac_key to do id match.

Reviewed By: jrodal98

Differential Revision: D31156925

fbshipit-source-id: a147de0936b1995d8946c75bb159cf1a82d64f73