Add script for deploying shared PCE resources (#120)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/120

PCE shared resources are shared across a single region. So we only need to manually run this script once per region and the resources will be shared across all advertisers in that region.
Note: PCE service library will have the logic to understand shared PCE resources and it will automatically use it on the publisher side

Next: Add a wiki page with instructions on how to deploy shared resources

Reviewed By: peking2

Differential Revision: D31322556

fbshipit-source-id: ae077b03a624386dde627ce07912d54f93a0967d