Remove the option to pass `protocol` to PrivateComputationService.id_match() because we already only use a hard-coded value (#103)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/103

I was talking with Jacob the other day and he pointed out to me that the protocol parameter passed in to the PrivateComputationService is actually ignored. In fact, we always just use UNION_PID because that's the only protocol supported: https://fburl.com/code/1qu6orva

This diff removes the option to pass protocol to PrivateComputationService.id_match().

In the future if more protocol is supported, we can explore other ways to pass the protocol parameter in. For example, passing in at PC instance creation time and save in the instance to be consumed by the id match stage.

Reviewed By: jrodal98

Differential Revision: D31209707

fbshipit-source-id: 43dcc33ec4d594d7cb96494c70d92c83af0a035d