complete undeploy functionality in partner-side `deploy.sh` (#147)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/147

**Context**

1. Undeploy feature in partner-side `deploy.sh` was originally designed to `undeploy` all AWS resources. It does not support undeploy sequentially (or partially) as it checks the existence of all required/expected tfstate files.
2. Undeploy logic was not complete. `terraform init` was missing before we run `terraform destroy`
*  what `terraform init` will do:

`terraform init` finds the correct backing tfstate file from s3 and populates a local .lock file with all that info.  Then if you run other commands, it knows which tfstate file you are referring to. (Note how you don't specify tfstate file path in plan, apply, destroy commands).)

* Why we need  `terraform init`  in `undeploy`:

If an advertiser successfully `deploy`ed all resources. And when the times comes for `undeploy`, we can't assume they will run it on the same container and that local .lock file will still exist. So if it invokes destroy directly, it won't know which tfstate file you are referring to.

**What**
In this diff, we
1. update `undeploy` logic: now `undeploy` command will destroy resources (PCE, VPC_peering, pce_shared, data ingestion) in sequence.
2. add terraform initialization before terraform destroy.

Reviewed By: peking2

Differential Revision: D31485156

fbshipit-source-id: 097a9cc5fef0e75e8a3618fc9af09d36cece9bf2