Update the terraform script to add tags for route table
Summary:
What:
1. Use default tag to setup the PCE id
2. Touch the default vpc route table to apply tags to route table.

Why:
The VPC create a default route table, so in tf script, we don't create a new route table. The default route table doesn't have the PCE id tag because it's not created by terraform. In the diff, we touch the default route table, so it will apply tags to it.

Differential Revision: D31544460

fbshipit-source-id: 9d30a13af6e6650840ea9d45fec0229fe6b06e3d