Update the terraform script to add tags for route table (#155)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/155

What:
1. Use default tag to setup the PCE id
2. Touch the default vpc route table to apply tags to route table.

Why:
The VPC create a default route table, so in tf script, we don't create a new route table. The default route table doesn't have the PCE id tag because it's not created by terraform. In the diff, we touch the default route table, so it will apply tags to it.

Reviewed By: peking2, ajaybhargavb, marksliva

Differential Revision: D31544460

fbshipit-source-id: c5b911b5dbb6b823f77a3c7ca8319ee41271afb9