de-dup private computation service wrapper functions, the easy part (#160)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/160

This stack focuses on consolidating the PL and PA coordinators into one single CLI.

This diff does 2 things (and some functions/variables renaming that should be self-explanatory)
1. Move pl_service_wrapper.py to private_computation_service_wrapper.py because it contains wrapper functions of pcs, not pl service anymore (which doesn't exist anymore);
2. Remove the functions from pa_coordinator.py that are complete duplicates of functions in private_computation_service_wrapper.py

Next, I'll move the remaining functions in pa_coordinator.py to private_computation_service_wrapper.py. Intentionally not doing it here so that I'm not over-complicating this diff; the remaining functions are not complete duplicates, so I cannot just delete them like how I did with those in this diff.

Reviewed By: jrodal98

Differential Revision: D31585359

fbshipit-source-id: 5973420c6659f57f916eff05b9e4c279494b78e8