remove instance repo update in run_stage (#162)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/162

## What

Remove an unneccessary update to the instance repository in run_stage

## Why

I noticed while working on another diff stack that this line could theoretically lead to a race condition. If we update the status in the instance repository to XXX_STARTED, then the instance monitor will start updating the instance. Since we update the instance again 10 lines down, it's possible they could do the update at the same time.

Thus, we should update the status in the instance repository only once. The update at the bottom will catch exceptions and set the status to failed, so it makes the most sense to update it here. There are unit tests to simulate this.

Reviewed By: Olivia-liu

Differential Revision: D31591182

fbshipit-source-id: 74f486c792e46cc7f216adff2a69d20e4ca9e063