Fix problem with GET requests not getting output at the end of the program (#164)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/164

GET deployment requests to the server could miss the last part of a command issued to them due to a race condition on the thread consuming the logs vs the process terminating.
Adding a second point of log consumption AFTER the process terminates guarantees that the entire log gets consumed.

Differential Revision: D31496041

fbshipit-source-id: 74dcea437a80b52f5c51d8a217eadc5e6d4b81db