copyright header for new dockerfile (#177)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/177

Got a high-pri task today saying that some file in the fbpcs repo is missing the copyright header.

I went to the [automated check up](https://www.internalfb.com/intern/opensource/github/repo/232509348613701/checkup) page and clicked the "Run Automated Checkup" green button, got an email with this paste P462617773, which indicates that the following file doesn't have the header.

Linter should have caught it (linter rule added in D30141706), but I guess it didn't work because this file is in the public_tld folder and we did some complicated mapping in the shipit config for this folder. I don't have time to look into it today, so leave that as a task.

Reviewed By: ajaybhargavb, jrodal98

Differential Revision: D31659591

fbshipit-source-id: a019feac8059d7ac18b95f88096641a3eac57261