Decoupling - Add logic in pa-coordinator to support decoupling.
Summary:
Changes:
Added logic to pa-coordinator for the new decoupling flow. Added a parameter is_decoupled to create_instance stage, based on which we will decide whether to call the new - decoupled flow or the legacyFlow.
Added a check to shard_aggregation stage, which will check the flow and accordingly will either take the decoupled_aggregation output or the computation output as the input path.

Next:
1. Add thrift support.
2. Update E2E tests.

Reviewed By: jrodal98

Differential Revision: D31750901

fbshipit-source-id: c934eace5d2926da37bd57786da8601c0f4886b1