Added method get_flow_cls_name to pc_instance. (#278)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/278

**Changes:**
Changes in this diff are
1. Added a method get_flow_cls_name is private_computation_instance.
2. Added a call to the method in shard aggregation service.

**Why:**
In PA we have two flows now - Decoupled Stage flow and Legacy flow. In the decoupled flow, instead of compute stage, we are calling decoupled attribution and decoupled aggregation stage.
Thus when we get to shard aggregator stage, we need to verify whether we arrived from decoupled_aggregation stage or compute stage so that we could pass the proper input path (compute output or decoupled_aggregation output).
In order to check that, we will check which flow we are using to determine the path.

Reviewed By: jrodal98

Differential Revision: D32020458

fbshipit-source-id: d6659fe87148e8390c0ef1ee852d93b89557cb7b