Decoupling - adding decoupled games to on-call scripts.
Summary:
**Changes:**
Modified on-call scripts and related files (CMakeLists.txt, Dockerfile.ubuntu) for new decoupled games.

**Why:**
We use on-call scripts to promote binaries in master to prod on a weekly basis.
The process is as follows
1. build the binaries and export them to AWS S3 (with rc tag).
2. Run E2E tests on the rc tagged binaries.
3. If tests passed, promote the binaries to prod.

Thus, once this diff is landed, decoupled_attribution and decoupled_aggregation binaries will be promoted to prod along with other binaries.

Reviewed By: gorel

Differential Revision: D31961557

fbshipit-source-id: 51bef6f4dea9a22251fcccb479a0d12985ded16f