Handle us-east-1 bucket creation separately in deploy.sh
Summary:
We've been met this issue for several times: https://fb.workplace.com/groups/331044242148818/posts/378945420692033/

The bucket can't be created in us-east-1 with deploy.sh.

It seems the problem is us-east-1 is the default region, and it doesn't need to specify `--create-bucket-configuration LocationConstraint="$region"`. While other regions need.

The solution is to not set `LocationConstraint` when it's us-east-1.

Reviewed By: anthonyzhang25, marksliva

Differential Revision: D32842279

fbshipit-source-id: b9b5dac55d2df52f42f9de09d3784214e7398cd2