Handle us-east-1 bucket creation separately in deploy.sh (#436)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/436

We've been met this issue for several times: https://fb.workplace.com/groups/331044242148818/posts/378945420692033/

The bucket can't be created in us-east-1 with deploy.sh.

It seems the problem is us-east-1 is the default region, and it doesn't need to specify `--create-bucket-configuration LocationConstraint="$region"`. While other regions need.

The solution is to not set `LocationConstraint` when it's us-east-1.

Reviewed By: taoyong-ty, anthonyzhang25, marksliva

Differential Revision: D32842279

fbshipit-source-id: 180a2159263999c911a6659dc0bde580ff4941e9