Remove emp::batcher
Summary:
The latest version of emp doesn't have `emp::batcher`. It was used as a way to batch the transfer of data between parties and reduce network usage, but it seems that the latest version of emp handles that batching automatically.

This diff removes `emp::batcher` from the codebase. The next diff will upgrade the version of emp that's used in fbcode.

This looks like a lot of changes but it's mostly the same type of change in each file.

For each class O:
- Remove `O::bool_size()`, `O::bool_data()`, and `O(int32_t len, const emp::block* b)`. These are all the methods that are required to support batcher.
- Add a new constructor, `O(T, int party)`, where T is the "non-private" version of O.

NOTE: These are the same changes I made in D30432552 and D30433391 the last time I tried upgrading the version of emp.

Reviewed By: gorel

Differential Revision: D32934300

fbshipit-source-id: 756320c7fda0d2758d7ae145c9ae75159b8d2690