Remove Timestamp class (#485)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/485

The latest version of emp doesn't really support integers that aren't 32 or 64 bits very well. See the [reveal](https://github.com/emp-toolkit/emp-tool/blob/master/emp-tool/circuits/integer.hpp#L183) functions which assume 32 or 64 bits. Revealing a negative number that uses a different number of bits will return an incorrect positive number.

The upshot is that the Timestamp class, which dynamically calculates the number of bits it needs, could break with the newest version of emp. I think the simplest thing to do is just remove it, for these reasons:
- The bit reduction has not yet been turned on. Right now timestamps still use 64 bits.
- I think using a fixed number of bits (i.e. 32 or even 17) would make the code simpler, and would give us a big gain already. The benefit we'd get from using 8 bits, say, is probably not worth the complexity here.
- Running attribution with minute or hour granularity would be a product decision, and I'm not sure if we're at a point where we can try to sell advertisers on a lossy computation.
-  PCF 2.0 doesn't support this bit reduction optimization anyway.

Reviewed By: ashiquemfb

Differential Revision: D32970874

fbshipit-source-id: a6b9a128f46ace63e8241f72f3ee2837ac182433