Add last click 2-7d attribution rule (#496)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/496

As title suggests, we want to support a longer attribution window. This diff does both the correctness testing as well as the actual implementation, since it is important to have them together

**Attribution logic**
We define 2 thresholds, a 1 day threshold and a 7 day threshold. The given touchpoint is attributed to the conversion if 3 things are true: it is a valid click, it is greater than the 1 day threshold, and it is less than the 7 day threshold. The structure of the code is simply copied from the 1d and 28d rules that are in the same file.

**Correctness Tests**
See the table for "LAST CLICK 2-7D" in this quip: https://fb.quip.com/IUHDApxKEAli. The first column explains the test in English.

Differential Revision: D33101914

fbshipit-source-id: 2115100f5d051cdb999177fb08a6d5ddc19166ce