Removing legacy attribution game reference from co-ordinator code.
Summary:
# Context
Decoupled attribution is now the default flow in production. We have done 5 successful pilots using the same. Thus at this time, it's safe to remove legacy attribution code. In this stack removing all references to the legacy attribution code.

# This Diff
In this diff, removing references to legacy attribution game in co-ordinator code.

Differential Revision: D33438464

fbshipit-source-id: e51ee119c60690aad3735cf4cd1732955d2dae11