Create `IFileUploadUtility` and its implementations to prepare for multi-cloud extension (#514)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/514

This stack is based on the discussions and decisions made in D33156029.

This purpose of the stack is to unblock GCP e2e runner. Eventually these classes in this diff will be replaced with similar ones in fbpcf. The fbpcf work and migration work will be tracked in GCP work stream.

The stack (short-term plan) will contain 3 diffs:

1/n: create a new interface, a S3 implementation, and a local implementation based on D33156029 (and unit tests).
2/n: Migrate `LifeIdSpineFileCombiner`, `UnionPIDDataPreparer` and `GenericSharder` off the current implementation to use the polymorphic `private_lift::file_upload_utils::upload()`.
3/n: Add a GCS implementation.

The medium-term migration will be:
1/n: FBPCF to support writing a `src` file to a given `dest` (in FileManagerUtil and IFIleManager classes).
2/n:  Migrate `LifeIdSpineFileCombiner`, `UnionPIDDataPreparer` and `GenericSharder` to use FBPCF. Remove unused classes from FBPCS.

Reviewed By: gorel

Differential Revision: D33356933

fbshipit-source-id: 25ca91d76a15cf788bca585fdf15783e57ceba34