Removing legacy attribution game reference from co-ordinator code. (#516)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/516

# Context
Decoupled attribution is now the default flow in production. We have done 5 successful pilots using the same. Thus at this time, it's safe to remove legacy attribution code. In this stack removing all references to the legacy attribution code.

# This Diff
In this diff, removing references to legacy attribution game in co-ordinator code.

Reviewed By: jrodal98

Differential Revision: D33438464

fbshipit-source-id: e0f4f44ab59e173107f0550109a614e738197647