Decoupling - Ad Id Compression on Aggregation. (#520)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/520

# Context

In Diff D32621605 we added Ad ID compression for PCF 2.0 aggregation game. As there is still some time for PCF 2.0 framework as well as games to launch, adding the same optimization to existing production game. So that we achieve more performance gains in upcoming pilots.

Please check the summary of D32621605 for detailed analysis on the functionality.

Reviewed By: chualynn

Differential Revision: D33539669

fbshipit-source-id: f76fbf74ee764e6be3acc52b10374149fa9a9a7f