Removing legacy attribution game reference from on call script. (#530)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/530

# Context
Decoupled attribution is now the default flow in production. We have done 5 successful pilots using the same. Thus at this time, it's safe to remove legacy attribution code. In this stack removing all references to the legacy attribution code.

# This Diff
In this diff, removing references to legacy attribution game in on call code used for building and promoting binaries to production.

Differential Revision: D33439178

fbshipit-source-id: 0fbf417349b48e0ad156629be5e338ef882457eb