Removing legacy attribution game reference from on call script. (#530)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/530

# Context
Decoupled attribution is now the default flow in production. We have done 5 successful pilots using the same. Thus at this time, it's safe to remove legacy attribution code. In this stack removing all references to the legacy attribution code.

# This Diff
In this diff, removing references to legacy attribution game in on call code used for building and promoting binaries to production.

Reviewed By: corbantek, chualynn

Differential Revision: D33439178

fbshipit-source-id: 7399ab2b5b9319c5266d981a09fb8cc424cf168f