Removing Legacy attribution code from fbpcs (#531)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/531

# Context
Decoupled attribution is now the default flow in production. We have done 5 successful pilots using the same. Thus at this time, it's safe to remove legacy attribution code. In this stack removing all references to the legacy attribution code.

# This Diff

In this diff, removing the actual attribution binary code from fbpcs. Some of the files in attribution folder are used by shard aggregation game and cannot be removed.

Reviewed By: ashiquemfb

Differential Revision: D33434221

fbshipit-source-id: ec3c4bf06bbe4391c14c8989eebee47bc3289ecc