correcting vector reserve value to reduce memory consumption. (#543)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/543

# Change
In this diff changed the variable passed to paddedArrays.reserve() function from numVals * maxArraySize to numVals.

# Analysis
The reserve function will allocate this memory for outer vector paddedArrays. Thus in this case, we end up allocating more than required memory for the outerArray.

A better way to use reserve in this case is

```
paddedArrays.reserve(numVals, new vector<int>(maxArrays))
```
However, if look in the method, we are creating paddedVec vector at line 156 and pushing data to it on line 159. So the resizing that we are trying to avoid in this case, is happening anyways in the inner method.

Thus only reserving numVal locations for the outer vector in this case.

Reviewed By: chualynn

Differential Revision: D33821873

fbshipit-source-id: cdf82d1750456386a60f4d6c08e691086ad5c5c2