add rc tag to testing docker image to fix build
Summary:
publish docker image job is broken because the tests fail because the task definition points at the old tag instead of the new tag.

We should update the task definition but because I don't know who the owner/what else that task definition is used for, I am meanwhile adding the tag its expecting back. We should remove this in the future once the task definition is updated.

Differential Revision: D34369800

fbshipit-source-id: 52d86a032eed1b88c77e1252556b57c70c54fe5a