Fix cloudwatch url retriever on publisher side issue (#604)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/604

On the publisher side, if the PCE is generated by using the PCE management tool, the log group will be using the shared log group.

Example: the container name is `onedocker-container-ee9bc805f22e40f9bbc107d5f006b6e1` while in the url, the log group name will contain `onedocker-container-shared-us-west-2`.

The fix is to replace the last part of the container name string with "-shared-<region>" if the last part matches regex (32 bit includes 0-9 and a-z).

This is hacky but the log_retriever method is already hacky :)

The edge case is that the advertiser is also using a 32bit string with 0-9 and a-z, but that should rarely happen.

Reviewed By: Olivia-liu

Differential Revision: D33592515

fbshipit-source-id: 46bc33d009f8f6bf051027807ac028ce21d1d567