Add dummy row when shard file size is 0 (#610)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/610

# Context
MARISA has been experiencing failure for more than 10 days as mentioned in [the post](https://fb.workplace.com/groups/331044242148818/posts/422172903035951). The links to all the investigation is summarized in [the quip](https://fb.quip.com/WraiAhmVOV8u).

As the result of investigation, we have found that empty shard data is causing the failure. We are able to reproduce the same failure on local PID protocol run with having server data empty.

# Details
As the temporary solution, we have determined to add dummy data for those empty shard data. In this diff, we are adding dummy data in the tmp file during PID prepare stage. Then, the tmp file with dummy data would be copied to actual S3 data file path to be used as the input for PID protocol.

Reviewed By: jrodal98

Differential Revision: D34437953

fbshipit-source-id: 476694740686ce0f45a491d061c52d5778fe5b91