Deprecate partial_container_retry_enabled field of PrivateComputationInstance
Summary:
**Background**
This field was added in D29509689 more than half a year ago when I worked on the *partial container failure recovery* feature. The work stream was deprioritized, and the feature is not usable. I'm cleaning it up to avoid confusion and poor maintenance. If we still want this feature, I would recommend redesigning it.

**This diff**
1. Intentionally not deleting `partial_container_retry_enabled` from `PrivateComputationInstance` class because from what I know, there's no safe way to delete a field from the instance class;
2. This diff is not as huge/scary as it looks. The main thing I did was removing this if branch
```
     if (
            ready_for_partial_container_retry(private_computation_instance)
            and not self._skip_partial_container_retry
        ):
```

Differential Revision: D34565572

fbshipit-source-id: 8a2dce8905948b989e2e9ec134ec6de0f8282eeb