Remove all fail_fast usages (#627)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/627

**Background**
The partial container failure recovery feature was only partially implemented more than half a year ago (see stack D28778025). The work stream was deprioritized, and the feature is not usable. I'm cleaning it up to avoid confusion and poor maintenance. If we still want this feature, I would recommend redesigning it.

**This diff**
`fail_fast` is saved in the instance and consumed by the id match stage. This diff removes all usages of it from the id match stage.

Differential Revision: D34566650

fbshipit-source-id: 64ce7d5a6e207cdc7c15e3c5080070e20467f71d