Change the SSE type of the data bucket to SSE-S3
Summary:
The current SSE type of the data bucket is SSE-KMS. It caused the ECS task role couldn't access it. We could either to attach a policy to the ECS task role or change the SSE type to SSE-S3. This diff is the later because on the publisher side we are also using SSE-S3 and we need to keep them consistent.

This diff made following changes:
1. Change the SSE type of the s3 bucket to SSE-S3.
2. Removed the server side encryption setting on the Kinesis because it is redundent since the S3 bucket is already has SSE enabled. (source: https://stackoverflow.com/questions/53752246/kinesis-firehose-kms-encryption)
3. Removed the kms policy from the Kinesis firehose role since it won't need it to access s3 bucket.
4. Removed the kms policy from the Glue crawler role since it won'te need it to access s3 bucket.

Differential Revision: D34605656

fbshipit-source-id: a23f3de75024f9c17f4bf18c4ec78bc3bc016672