Adding pcf2 aggregation service. (#650)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/650

# Context
With the Open sourcing of PCF 2.0, we will also be releasing attribution and aggregation based on PCF 2.0 to production. While promoting the PCF 2.0 games to production and making them the default games for production runs, we would also like to keep an option to run the current production games (i.e. attribution and aggregation game based on PCF 1.0) in prod for disaster recovery/fall back scenarios.

Thus in this stack of diffs
1. I will be adding a new flow for PCF 2.0 games in PCS. The new flow will be
    -  CREATED -> INPUT_DATA_VALIDATION -> ID_MATCH -> ID_MATCH_POST_PROCESS -> PREPARE -> PCF2_ATTRIBUTION -> PCF2_AGGREGATION -> AGGREGATE POST_PROCESSING_HANDLERS

2.Will be adding new service for PCF 2.0 attribution
3.Will be adding a new service for  PCF 2.0 aggregation.
4.Directory where PCF 2.0 attribution and aggregation binaries are placed would be -

```
private_attribution/prf2_attribution and private_attribution/pcf2_aggregation
 ```
5.Adding end to end tests for the new path.

# Technical decision
Leveraging the dynamic stage flow architecture in PCS to efficiently manage the stages without many changes that would break backward compatibility. While this approach does involve some temporary duplication of code, the duplicate code will be soon removed, as legacy attribution and aggregation games will be marked deprecated once we release PCF 2.0 to production and after 5 successful production runs (setting this number as we used the same number last half), we will remove the legacy code and flow.
This is least error prone way to manage different flows as will make it easier for us to fallback on the legacy flow in case of any errors/issues on new flow.

# This Diff
Adding new service to support PCF 2.0 aggregation as well as test cases for the same.

# This Stack
1. Specify path for PCF 2.0 binaries and add entries for them in private_computation_game.
2. Create service for pcf2 attribution.
3. Create service for pcf 2 aggregation.
4. Create new flow for PCF2.
5. Create additional test in fbpcs_e2e_runner.
6. Create additional test in automated_e2e_runner.

Differential Revision: D34658666

fbshipit-source-id: 364a6222ec6e24fa75b1b5b848ba0b577912d813