fix the bugs in minimum permission policy (#676)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/676

In computation UI, we need the user to provide the aws_access_key_id and aws_secret_access_key. Usually it's the admin credentials, but for the computation UI project, we have identified the policy with minimum permissions to run the calculation. This diff is to fix a few permissions.
1. We're using the wrong database name. It should be "mpc-events-db-<TAG>" instead of "fb_pc_data_<TAG>".
2. The "fb_pc_data_<TAG>" should be the table name.
3. Missed one resource for the GLUE permission.

Reviewed By: marksliva

Differential Revision: D34773691

fbshipit-source-id: 43c340b79e13b3ba78d37f634816faf62ba73be8