remove gtest from pcf2 attribution
Summary:
In pcf2 attribution we are using gtest to provide access to unit testing. We are now adding pcf2 binaries to oncall scripts and in the script we exclude the test files during build and release.
Thus in diff, removing gtest from AttributionGame, so that we do not have to include any test specific dependency in build scripts.

Differential Revision: D34947330

fbshipit-source-id: 41687a2c1dadcabb3d400dfcf3d3584ed0003196