remove gtest from pcf2 attribution (#707)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/707

In pcf2 attribution we are using gtest to provide access to unit testing. We are now adding pcf2 binaries to oncall scripts and in the script we exclude the test files during build and release.
Thus in diff, removing gtest from AttributionGame, so that we do not have to include any test specific dependency in build scripts.

Reviewed By: chualynn

Differential Revision: D34947330

fbshipit-source-id: 3aad7d277635241cf4913b38310037f1f4da0879