add data field quality threshold validation
Summary:
The data validation threshold config is now used at the end of the data
validator to determine if a file has enough usable data in it according to the
thresholds. This validation threshold config will default when not specified.

The value field threshold is only applied to the number of rows that are not
empty.

Differential Revision: D34984488

fbshipit-source-id: 5335bbcb0baf3d51418de4c786d2666810ce4945