Correcting path of onedocker binary (#740)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/740

Initially for PCF2 binary, I added - pcf2_attribution and pcf2_aggregation as the location. However, I see that the rest of the binaries are located in the S3 structure - private_attribution/{binary_name}.
Having just pcf2_attribution and pcf2_aggregation was causing an issue in onedocker-runney - packageParser. Created a fix for the same - D35087015, but change for pcf2 to match the general pattern.

Differential Revision: D35059153

fbshipit-source-id: 7ef230cb4c5a36cc4d7b6ed37ae19c9ce1c3340c