Added transformation of file/S3 paths in CLI; and improved unit testing for create_instance
Summary:
Added transformation of file/S3 paths in CLI. When the path is not a file that exists on the system, we attempt to convert the path to an S3 path in the “virtual-hosted-style access” format as suggested here T114608069.
Made improvements to the unit testing for `create_instance`, `run_instance`, `run_instances`, `run_study` by using actual temporary files and temporary folders rather than hardcoded file and folder names.

Reviewed By: jrodal98

Differential Revision: D35087543

fbshipit-source-id: c71a9a68b46b246b4b2db20a7a2e99321b3b7a7b