Adding support for test repository in upload-binaries-to-s3
Summary:
# Context

All the private computation production binaries as stored in a public S3 Repository - one-docker-repository-prod. Due to security concerns as well as not to interfere with with the production code currently run by advertisers, we cannot upload binaries created on unreleased code to this public repo.

This creates a challenge in order to test the binaries with PCS service.
In order to resolve this issue, in this stack I am adding support for running custom binaries through PCS and thus using end to end tests and EP.

For this, I have created a separate private S3 bucket - one-docker-repository-test. We can upload any unreleased code to this buckets as this are private and separate from production route. In this stack of diffs I will be adding support for using one-docker-repository-test through PCS as well as build scripts.

# This Diff
In this diff, we are modifying build-binaries.sh and related to scripts to be now used solely for building fbpcs from devserver and uploading the same to private S3 repository for test purposes.
This binaries will be uploaded under the tag of the USER uploading the script.

# This Stack
1.  Add support for uploading to one-docker-repository-test in build script.
2. Add support for one-docker-repository-test in PCS for MPC binaries.
3. Add support for one-docker-repository-test in PCS for PID binaries.
4. Add option to fbpcs_e2e_runner to use test repository.
5. Add path to EP for custom repository.
6.  (Not a diff but) - Publish documentation/post to detail use of the new functionality, as well as guidelines for developers to use test their custom binaries.

Reviewed By: gorel, jrodal98

Differential Revision: D35106690

fbshipit-source-id: b2ec39ece9a7242cfe2c45561b0053cccb93fdda