Change the behavior of BitString (#149)
Summary:
X-link: https://github.com/facebookresearch/fbpcf/pull/149

Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/758

As title.
Current BitString type have some inconsistent behavior.

For example, when initialized by a batch of `std::vector<bool>`, it is expecting {the vector of first bits in all vecs, the vector of second bits in all vecs...} instead of "the first vector, the second vector". This doesn't agree with other batched types.

This diff changes this behavior to align with other batch types.

Reviewed By: elliottlawrence

Differential Revision: D35122217

fbshipit-source-id: dd743e3562146cecc81db9c68819e072eff2235b