Change the behavior of BitString
Summary:
As title.
Current BitString type have some inconsistent behavior.

For example, when initialized by a batch of `std::vector<bool>`, it is expecting {the vector of first bits in all vecs, the vector of second bits in all vecs...} instead of "the first vector, the second vector". This doesn't agree with other batched types.

This diff changes this behavior to align with other batch types.

Differential Revision: D35122217

fbshipit-source-id: 197b33f3e01e9f2d0a2cff81eab06e5f65db1d80