Add a new GCP PCE deployment docker image.
Summary:
Context:
In order to deploy PCEs in GCP, we need a Docker image that includes the Google Cloud SDK. After talking this through with taoyong-ty and zhuang-93, we decided that a new image was the best route for right now. It may be merged with the AWS later down the road.

This Diff:
This diff creates a new docker file that will be used to create a pce deployment image for GCP. It also updates the existing build docker script to support a new flag that tells it to build the gcp image rather than the aws image.

Differential Revision: D35114991

fbshipit-source-id: 2ea78bf04334ee720f3fd0feab74d0eb19730c74